// src/models/importedMachines.ts
/**
 * Ravimoshark Contract API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: dev@ravimoshark.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Machine imported from SAGE.
 */
export interface ImportedMachines {
    id?: number;
    identification?: string;
    MachineId?: number;
    reviewed?: boolean;
    refContract?: string;
}